+++ packages/elements/src/__mocks__/use-resize-observer.ts
@@ -1,0 +1,3 @@
+const useResizeObserver = jest.requireActual('use-resize-observer');
+
+export default useResizeObserver;