+++ packages/elements/src/index.ts
@@ -5,3 +5,4 @@ export * from './containers/TableOfContents';
 export * from './containers/TryIt';
 export * from './containers/StoplightProject';
 export * from './containers/APIComponent';
+export { SchemaViewer, ISchemaViewerProps } from './components/SchemaViewer';